Update : 사용자 수정 및 삭제 기능 업데이트
자신이 작성한 게시글만 수정하거나 삭제할 수 있도록 업데이트 하였다.

+++ src/main/java/kr/codesqaud/cafe/controller/ArticleController.java
@@ -57,8 +57,14 @@ public class ArticleController {
 
     @GetMapping("/updateForm/{articleId}")
     public String update(@PathVariable int articleId,
+                         @SessionAttribute(name = ConstConfig.SESSION_ID) User user,
                          Model model) {
+        // TODO : 검증 공부 후 다시 리펙토링 할 예정
         Article article = articleRepository.findByArticleId(articleId);
+        if (!article.getWriter().equals(user.getUserId())) {
+            return "redirect:/qna/list";
+        }
+
         model.addAttribute(article);
         return "qna/updateForm";
     }
@@ -72,8 +78,13 @@ public class ArticleController {
     }
 
     @DeleteMapping("/delete/{articleId}")
-    public String delete(@PathVariable int articleId) {
-        articleRepository.delete(articleId);
+    public String delete(@PathVariable int articleId,
+                         @SessionAttribute(name = ConstConfig.SESSION_ID) User user) {
+        Article article = articleRepository.findByArticleId(articleId);
+        // TODO : 검증 공부 후 다시 리펙토링 할 예정
+        if (article.getWriter().equals(user.getUserId())) {
+            articleRepository.delete(articleId);
+        }
         return "redirect:/qna/list";
     }
